# Speed up windows mapping