@clefebvre 2018-02-02
With the addition of https://github.com/linuxmint/muffin/pull/308, we go down from 12s to about 7s.