@clefebvre 2018-02-02
With the timeout_add to create the window list buttons, we're now down to 6s for defocus/map, and the creation delay goes from 4s to about 300ms.